pyproject.toml
[project]
name = "chat-app-llm"
version = "0.1.0"
description = "Chat-GPT Clone with Groq LLM and Pinecone Memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "groq>=0.11.0",
    "pinecone>=5.0.0",
    "sentence-transformers>=2.2.0",
]
